fix(daemon): preserve lightweight Bun startup
Amp-Thread-ID: https://ampcode.com/threads/T-01a037b7-66db-71f0-91e7-1578b383afb2
